Initial commit
feat: New app project created with the name "Restart".

feat: "Sounds" folder created and 2 sound files added to it.

feat: Added an image, as an app icon, to the Assets folder.

feat: Added a new Blue color, as the  Accent color - #53A4D6, to the Assets folder.

feat: Created a "Characters" folder in the Assets folder and added to character images to it.

feat: Created a colorset "Colors" folder in the Assets folder and add 2 colors to it. Blue - #53A4D6 and Red - #C67976.